introduce `AbstractActiveThreadPoolRunner`
